+++ LibraryAppProject/Styles/ColorStyles.swift
@@ -1,18 +1,5 @@
 import SwiftUI
 
-
-var customFont: Font {
-    return .custom("Helvetica Neue", size: 16)
-}
-
-var serifShapedFont: Font {
-    return .custom("Times New Roman", size: 20)
-}
-
-var subsectionFont: Font {
-    return .custom("Helvetica Neue", size: 20)
-}
-
 var customColor: Color {
     return Color(UIColor(red: 106 / 255, green: 48 / 255, blue: 72 / 255, alpha: 1))
 }
