+++ LibraryAppProject/Styles/FontStyles.swift
@@ -1,0 +1,13 @@
+import SwiftUI
+
+var customFont: Font {
+    return .custom("Helvetica Neue", size: 16)
+}
+
+var serifShapedFont: Font {
+    return .custom("Times New Roman", size: 20)
+}
+
+var subsectionFont: Font {
+    return .custom("Helvetica Neue", size: 20)
+}
